refactor: Update MCP types
- Consolidate MCP types by moving McpServerState and McpServerStatus to mcp.types.ts
- Update MCPSection and McpManager to utilize the new enum for server status
