+++ src/components/settings/sections/MCPSection.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from 'react'
 
 import { useSettings } from '../../../contexts/settings-context'
 import SmartComposerPlugin from '../../../main'
-import { McpTool } from '../../../types/mcp.types'
-import { McpManager, McpServerState } from '../../../utils/mcp/mcpManager'
+import { McpServerState, McpTool } from '../../../types/mcp.types'
+import { McpManager } from '../../../utils/mcp/mcpManager'
 import { ObsidianButton } from '../../common/ObsidianButton'
 import { ObsidianToggle } from '../../common/ObsidianToggle'
 import { AddMcpServerModal, EditMcpServerModal } from '../McpServerModal'
